requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
patchright==1.58.0
google-cloud-storage==2.18.0
google-auth==2.36.0
pydantic-settings==2.7.0
slowapi==0.1.9
python-dotenv==1.0.1
